Add Multi Project
Create Console App
