+++ .gitignore
@@ -1,2 +1,3 @@
-*/build/
-*/.gradle/
+**/build/
+**/.gradle/
+**/.idea/
