+++ console/build.gradle.kts
@@ -19,6 +19,8 @@ repositories {
 }
 
 dependencies {
+    implementation(project(":editor"))
+
     implementation("org.apache.avro:avro:1.11.0")
     implementation("org.apache.avro:avro-compiler:1.11.0")
     implementation("org.slf4j:slf4j-api:1.7.32")
