+++ console/src/main/kotlin/editor/Constants.kt
@@ -1,0 +1,9 @@
+package editor
+
+class Constants {
+    companion object {
+        const val ENV_KAFKA_SERVER = "KAFKA_SERVER"
+        const val ENV_SCHEMA_REGISTRY_SERVER = "SCHEMA_REGISTRY_SERVER"
+        const val ENV_OUTPUT_PATH = "EDITOR_OUTPUT_PATH"
+    }
+}
