--- editor/src/main/kotlin/editor/schemaregistry/console/SchemaRegistryConsole.kt
@@ -1,23 +1,0 @@
-package editor.schemaregistry.console
-
-import editor.schemaregistry.SchemaRegistry
-import editor.schemaregistry.SubjectSchema
-
-class SchemaRegistryConsole(
-    val schemaRegistry: SchemaRegistry = SchemaRegistry()
-) {
-    fun list(topic: String?) {
-        println("Listing Schemas")
-
-        if (topic != null)
-            printSchema(schemaRegistry.getSchema(topic))
-        else
-            SchemaRegistry().getAllSchemas()
-                .forEach {
-                    printSchema(it)
-                }
-    }
-
-    private fun printSchema(subjectSchema: SubjectSchema) =
-        println("${subjectSchema.topic}: ${subjectSchema.schema}")
-}
