--- libexample/.gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Sat Dec 11 18:33:46 BRT 2021
-gradle.version=7.1
