+++ settings.gradle.kts
@@ -1,0 +1,2 @@
+rootProject.name = "kafka-avro-editor"
+include("editor", "console")
